feat: error 处理

+++ internal/api/request/cart_request.go
@@ -1,6 +1,6 @@
 package request
 
 type AddToCartRequest struct {
-	ProductID uint `json:"product_id" binding:"required"`
-	Quantity  int  `json:"quantity" binding:"required"`
+	ProductID uint64 `json:"product_id" binding:"required"`
+	Quantity  int    `json:"quantity" binding:"required"`
 }
